Added redis metrics to subscription set
